requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
langsmith>=0.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
psycopg2-binary>=2.9.0